+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## [2.0.0](https://github.com/sketch7/ngx.ux/compare/1.2.1...2.0.0) (2020-xx-xx)
+
+### Features
+
+- **deps:** update dependencies to support Angular v10
+- **deps:** add dependency on `tslib: ^2.0.0`
+- **viewport:** add `size$` to `ViewportService`
+
+### Bug Fixes
+
+- **viewport:** throw error when invalid size type is used in matcher
+
+### BREAKING CHANGES
+
+- **viewport:** make several utils/models internal only
+- **viewport:** change `ViewportSizeType` and `UX_VIEWPORT_DEFAULT_BREAKPOINTS` has been changed (see readme)
+- **viewport:** remove `ViewportService.calculateItemsPerRow` function
+
 ## [1.2.1](https://github.com/sketch7/ngx.ux/compare/1.1.2...1.2.0) (2020-10-28)
 
 ### Bug Fixes